requirements.txt
pandas
numpy
matplotlib
scikit-learn
ultralytics
torch
Pillow
requests
uvicorn
streamlit
onnxruntime